ai-backend/requirements.txt
fastapi
uvicorn
torch
Pillow
python-multipart
opencv-python
pandas